:truck: Move `credentials.api.ts` outside /public
